pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evalinfra-sdk"
version = "0.1.0"
description = "Python SDK for LLM evaluation infrastructure"
authors = [{name = "Eval Infra Team"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    # Core LLM dependency
    "litellm",
    
    # Data and validation
    "pydantic>=2.0.0",
    
    # Async HTTP
    "aiohttp",
    
    # Document parsing
    "reductoai",
    
    # Utilities
    "python-dotenv",
    "tqdm",
    "ijson",
]

[project.optional-dependencies]
# Testing and development
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]
namespaces = false

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
